Break import cycle at util/state.js -> components/form/user-settings.js
Move i18n helper function out of user-settings.js.
