+++ lib/util/state.js
@@ -12,7 +12,7 @@ import React from 'react'
 import styled from 'styled-components'
 
 import { MainPanelContent } from '../actions/ui-constants'
-import { summarizeQuery } from '../components/form/user-settings'
+import { summarizeQuery } from '../components/form/user-settings-i18n'
 import FormattedMode from '../components/util/formatted-mode'
 import Strong from '../components/util/strong-text'
